feat: Put helper functions in a file
